Removed AbstractBipartiteGraph
