+++ src/main/java/edu/ut/ece/social/graph/BipartiteGraph.java
@@ -1,34 +1,92 @@
 package edu.ut.ece.social.graph;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.graph.*;
+import com.google.common.graph.Graph;
+import com.google.common.graph.MutableValueGraph;
+import com.google.common.graph.ValueGraph;
+import com.google.common.graph.ValueGraphBuilder;
+import com.google.common.graph.EndpointPair;
+import com.google.common.graph.ElementOrder;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.*;
 
+import static com.google.common.base.Preconditions.checkArgument;
+
 /**
  * A weighted, undirected bipartite graph.
  */
-public final class BipartiteGraph<N, V> extends AbstractBipartiteGraph<N> implements ValueGraph<N, V> {
+public final class BipartiteGraph<N, V> implements ValueGraph<N, V> {
 
     private final MutableValueGraph<N, V> underlyingGraph = ValueGraphBuilder.undirected().build();
+    private final Set<N> leftSideNodes;
+    private final Set<N> rightSideNodes;
 
-    protected BipartiteGraph() {
+    BipartiteGraph() {
         this(new HashSet<>(), new HashSet<>());
     }
 
     private BipartiteGraph(Set<N> leftSideNodes, Set<N> rightSideNodes) {
-        super(leftSideNodes, rightSideNodes);
+        this.leftSideNodes = leftSideNodes;
+        this.rightSideNodes = rightSideNodes;
+        leftSideNodes.forEach(this::addLeftSideNode);
+        rightSideNodes.forEach(this::addRightSideNode);
     }
 
-    @Override
+    public boolean addLeftSideNode(N node) {
+        return addNode(node, leftSideNodes);
+    }
+
+    public boolean addRightSideNode(N node) {
+        return addNode(node, rightSideNodes);
+    }
+
+    public boolean isLeftSideNode(N node) {
+        return leftSideNodes.contains(node);
+    }
+
+    public boolean isRightSideNode(N node) {
+        return rightSideNodes.contains(node);
+    }
+
+    public ImmutableSet<N> leftSideNodes() {
+        return ImmutableSet.copyOf(leftSideNodes);
+    }
+
+    public ImmutableSet<N> rightSideNodes() {
+        return ImmutableSet.copyOf(rightSideNodes);
+    }
+
+    boolean addNode(N node, Set<N> nodeSet) {
+        if (isLeftSideNode(node) || isRightSideNode(node)) {
+            return false;
+        }
+
+        nodeSet.add(node);
+        return addNodeToUnderlyingGraph(node);
+    }
+
+    public boolean removeNode(N node) {
+        leftSideNodes.remove(node);
+        rightSideNodes.remove(node);
+        return removeNodeFromUnderlyingGraph(node);
+    }
+
+    protected void addEdgeSetup(N leftSideNode, N rightSideNode) {
+        checkArgument(!isRightSideNode(leftSideNode),
+                "Provided left side node is already a right side node");
+        checkArgument(!isLeftSideNode(rightSideNode),
+                "Provided right side node is already a left side node");
+
+        leftSideNodes.add(leftSideNode);
+        rightSideNodes.add(rightSideNode);
+    }
+
     protected boolean addNodeToUnderlyingGraph(N node) {
         return underlyingGraph.addNode(node);
     }
 
-    @Override
     protected boolean removeNodeFromUnderlyingGraph(N node) {
         return underlyingGraph.removeNode(node);
     }
@@ -154,5 +212,3 @@ public final class BipartiteGraph<N, V> extends AbstractBipartiteGraph<N> implem
         return underlyingGraph.hashCode();
     }
 }
-
-
